requirements.txt
Flask==3.1.2
gunicorn==23.0.0
matplotlib==3.10.6
networkx==3.4.2
openai==1.108.1
pydantic==2.11.9
python-dotenv==1.1.1
Requests==2.32.5
tiktoken==0.11.0
xai_sdk==1.2.0
